Create secrets.json
Replace <YOUR_ALCHEMY_API_KEY> with your Alchemy API key. You can sign up for an Alchemy account and obtain the API key from their website. Additionally, replace <YOUR_METAMASK_MNEMONIC> with the mnemonic phrase of your Metamask wallet. This is typically a 12 or 24-word phrase that you should keep secure and private.

Save the secrets.json file.
Remember to add the secrets.json file to your .gitignore file to ensure it is not committed to a public repository. It is important to keep sensitive information confidential and avoid exposing it to unauthorized individuals.